requirements.txt
streamlit
pydantic
pandas
requests
beautifulsoup4
scikit-learn
openai
python-dotenv
tiktoken
lxml
openpyxl
anthropic
google-generativeai
sentence-transformers
chromadb
pytest
